+++ src/ToySolver/SAT/PBO/BCD.hs
@@ -61,7 +61,7 @@ solve cxt solver = solveWBO (C.normalize cxt) solver
 
 solveWBO :: C.Context cxt => cxt -> SAT.Solver -> IO ()
 solveWBO cxt solver = do
-  SAT.setEnableBackwardSubsumptionRemoval solver True
+  SAT.modifyConfig solver $ \config -> config{ SAT.configEnableBackwardSubsumptionRemoval = True }
   ub <- atomically $ C.getSearchUpperBound cxt
   loop (IntSet.fromList [lit | (lit,_) <- sels], IntSet.empty) [] ub