+++ CHANGELOG.markdown
@@ -3,6 +3,7 @@
 * add experimental getAssumptionsImplications functions to ToySolver.SAT module.
 * add getFixedLiterals to ToySolver.SAT module.
 * use 'mwc-random' package instead of 'random' package.
+* introduce 'Config' data type in ToySolver.SAT to simplify configulation management.
 
 0.3.0
 -----